# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Associati digilab",
        "description": "Api per la gestione degli associati del digilab",
        "version": "0.1"
    },
    "paths": {
...
        "/api/nuova-iscrizione": {
            "post": {
                "tags": [
                    "gestione iscrizioni"
                ],
                "summary": "Aggiunge un nuovo iscritto ad un corso",
                "description": "inserisce i dati di un'iscrizione ad un corso'",
                "operationId": "2ba79567dfd5a108d24010c1d39e5408",
                "requestBody": {
                    "description": "dati iscrizione",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "associato_id",
                                    "corso_id",
                                    "data_iscrizione",
                                    "pagato"
                                ],
                                "properties": {
                                    "associato_id": {
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "corso_id": {
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "data_iscrizione": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "data_fine": {
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "costo": {
                                        "type": "string",
                                        "format": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "corso aggiunto"
                    },
                    "204": {
                        "description": "dati mancanti o non validi"
                    },
                    "500": {
                        "description": "Errore nell'inserimento del nuovo corso"
                    }
                }
            }
        },
        "/api/elimina-iscrizione/{id}": {
            "delete": {
                "tags": [
                    "gestione iscrizioni"
                ],
                "summary": "Elimina un 'iscrizione tramite id associato",
                "operationId": "cf4f5ccd1662b801625c386de15f1eab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id dell\\'associato da eliminare come iscrizione",
                        "required": true,
                        "schema": {
                            "description": "id associato",
                            "type": "string",
                            "minimum": 16
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "iscrizione eliminata"
                    },
                    "404": {
                        "description": "iscrizione non trovata"
                    },
                    "500": {
                        "description": "errore dei dati ricevuti"
                    }
                }
            }
        },
        "/api/iscrizioni-per-corso/{id}": {
            "get": {
                "tags": [
                    "gestione iscrizioni"
                ],
                "summary": "Recupera tutti gli associati iscritti ad un corso selezionato",
                "operationId": "f3781b8cce377b5fdd690b8226fcda33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del corso",
                        "required": true,
                        "schema": {
                            "description": "ID corso",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista delle iscrizioni",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "corso_id": {
                                                "type": "string",
                                                "format": "text"
                                            },
                                            "titolo": {
                                                "type": "string",
                                                "format": "text"
                                            },
                                            "nome": {
                                                "type": "string",
                                                "format": "text"
                                            },
                                            "cognome": {
                                                "type": "string",
                                                "format": "text"
                                            },
                                            "data_iscrizione": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "pagato": {
                                                "type": "boolean"
                                            },
                                            "associato_id": {
                                                "type": "string",
                                                "format": "text"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Dati mancanti"
                    },
                    "404": {
                        "description": "Corso non trovato"
                    },
                    "500": {
                        "description": "Errore nel recupero della lista degli associati"
                    }
                }
            }
        },
        "/api/iscrizioni-per-associato/{id}": {
            "get": {
                "tags": [
                    "gestione iscrizioni"
                ],
                "summary": "Recupera tutti i corsi ai quali è iscritto l'associato selezionato",
                "operationId": "3e12822a941946831553f562a927ced3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del'associato",
                        "required": true,
                        "schema": {
                            "description": "id associato",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista delle iscrizioni",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "corso_id": {
                                                "type": "string",
                                                "format": "text"
                                            },
                                            "titolo": {
                                                "type": "string",
                                                "format": "text"
                                            },
                                            "data_iscrizione": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "pagato": {
                                                "type": "boolean"
                                            },
                                            "associato_id": {
                                                "type": "string",
                                                "format": "text"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Dati mancanti"
                    },
                    "404": {
                        "description": "Associato non trovato"
                    },
                    "500": {
                        "description": "Errore nel recupero della lista dei corsi"
                    }
                }
            }
        }
    },
...
}